Move utils.py into utils folder
Added it to __init__ so that the functions there can still be
imported directly.

The passwords as well are in the utils folder now.
